refactor(#0102): split application files into one-file-one-class model
- Remove monolithic models.py, views.py, and forms.py files
- Split models into individual files in models/ directory
- Split views into logical modules in views/ directory:
  - dashboard.py: dashboard view
  - expense.py: expense CRUD operations
  - month.py: month management
  - payment.py: payment recording and editing
  - payee.py: payee management
  - budget.py: budget operations
  - payment_method.py: payment method management
- Split forms into focused modules in forms/ directory:
  - expense.py: ExpenseForm with complex validation
  - payment.py: PaymentForm and ExpenseItemEditForm
  - payee.py: PayeeForm
  - budget.py: BudgetForm
  - payment_method.py: PaymentMethodForm
- Maintain backward compatibility through __init__.py imports
- Update test mocking to use correct module paths
- All 179 tests continue to pass
