+++ expenses/forms/payment_method.py
@@ -1,0 +1,13 @@
+from django import forms
+from ..models import PaymentMethod
+
+
+class PaymentMethodForm(forms.ModelForm):
+    """Form for creating and editing payment methods"""
+    
+    class Meta:
+        model = PaymentMethod
+        fields = ['name']
+        widgets = {
+            'name': forms.TextInput(attrs={'class': 'form-control', 'placeholder': 'Enter payment method name'}),
+        }
